requirements.txt
fastapi
uvicorn
pydantic
PyMySQL
SQLAlchemy
python-dotenv 
psycopg2
psycopg2-binary 
